+++ src/js/plugins/libvorbis-plugin.js
@@ -15,6 +15,31 @@ const RecordEngine = videojs.getComponent('RecordEngine');
  */
 class LibVorbisEngine extends RecordEngine {
     /**
+     * Creates an instance of this class.
+     *
+     * @param  {Player} player
+     *         The `Player` that this class should be attached to.
+     *
+     * @param  {Object} [options]
+     *         The key/value store of player options.
+     */
+    constructor(player, options) {
+        super(player, options);
+
+        /**
+         * Enables console logging for debugging purposes.
+         *
+         * @type {boolean}
+         */
+        this.debug = false;
+        /**
+         * Specifies the sample rate to encode at.
+         *
+         * @type {number}
+         */
+        this.sampleRate = 32000;
+    }
+    /**
      * Setup recording engine.
      *
      * @param {LocalMediaStream} stream - Media stream to record.