+++ src/js/videojs.record.js
@@ -162,10 +162,13 @@ class Record extends Plugin {
         this.pictureInPicture = recordOptions.pip;
         this.recordTimeSlice = recordOptions.timeSlice;
         this.autoMuteDevice = recordOptions.autoMuteDevice;
+        this.pluginLibraryOptions = recordOptions.pluginLibraryOptions;
 
         // video/canvas settings
         this.videoFrameWidth = recordOptions.frameWidth;
         this.videoFrameHeight = recordOptions.frameHeight;
+        this.videoFrameRate = recordOptions.videoFrameRate;
+        this.videoBitRate = recordOptions.videoBitRate;
         this.videoEngine = recordOptions.videoEngine;
         this.videoRecorderType = recordOptions.videoRecorderType;
         this.videoMimeType = recordOptions.videoMimeType;
@@ -561,6 +564,8 @@ class Record extends Plugin {
             // listen for events
             this.engine.on(Event.RECORD_COMPLETE, this.engineStopCallback);
 
+            this.engine.pluginLibraryOptions = this.pluginLibraryOptions;
+
             // audio settings
             this.engine.bufferSize = this.audioBufferSize;
             this.engine.sampleRate = this.audioSampleRate;
@@ -582,6 +587,8 @@ class Record extends Plugin {
             // video/canvas settings
             this.engine.videoWorkerURL = this.videoWorkerURL;
             this.engine.videoWebAssemblyURL = this.videoWebAssemblyURL;
+            this.engine.videoBitRate = this.videoBitRate;
+            this.engine.videoFrameRate = this.videoFrameRate;
             this.engine.video = {
                 width: this.videoFrameWidth,
                 height: this.videoFrameHeight